third_party/blink/web_tests/external/wpt/html/dom/reflection-embedded-expected.txt

This is a testharness.js-based test.
Found 76 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] area.hreflang: typeof IDL attribute
  assert_equals: expected "string" but got "undefined"
[FAIL] area.hreflang: IDL get with DOM attribute unset
  assert_equals: expected (string) "" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to ""
  assert_equals: IDL get expected (string) "" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: IDL get expected (string) " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to undefined
  assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to 7
  assert_equals: IDL get expected (string) "7" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to 1.5
  assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to "5%"
  assert_equals: IDL get expected (string) "5%" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to "+100"
  assert_equals: IDL get expected (string) "+100" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to ".5"
  assert_equals: IDL get expected (string) ".5" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to true
  assert_equals: IDL get expected (string) "true" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to false
  assert_equals: IDL get expected (string) "false" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to object "[object Object]"
  assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to NaN
  assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to Infinity
  assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to -Infinity
  assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to "\\0"
  assert_equals: IDL get expected (string) "\\0" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to null
  assert_equals: IDL get expected (string) "null" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to object "test-toString"
  assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
[FAIL] area.hreflang: setAttribute() to object "test-valueOf"
  assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
[FAIL] area.hreflang: IDL set to ""
  assert_equals: getAttribute() expected "" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: getAttribute() expected " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got "test-valueOf"
[FAIL] area.hreflang: IDL set to undefined
  assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to 7
  assert_equals: getAttribute() expected "7" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to 1.5
  assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to "5%"
  assert_equals: getAttribute() expected "5%" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to "+100"
  assert_equals: getAttribute() expected "+100" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to ".5"
  assert_equals: getAttribute() expected ".5" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to true
  assert_equals: getAttribute() expected "true" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to false
  assert_equals: getAttribute() expected "false" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to object "[object Object]"
  assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to NaN
  assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to Infinity
  assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to -Infinity
  assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to "\\0"
  assert_equals: getAttribute() expected "\\0" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to null
  assert_equals: getAttribute() expected "null" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to object "test-toString"
  assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
[FAIL] area.hreflang: IDL set to object "test-valueOf"
  assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
[FAIL] area.type: typeof IDL attribute
  assert_equals: expected "string" but got "undefined"
[FAIL] area.type: IDL get with DOM attribute unset
  assert_equals: expected (string) "" but got (undefined) undefined
[FAIL] area.type: setAttribute() to ""
  assert_equals: IDL get expected (string) "" but got (undefined) undefined
[FAIL] area.type: setAttribute() to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: IDL get expected (string) " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got (undefined) undefined
[FAIL] area.type: setAttribute() to undefined
  assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
[FAIL] area.type: setAttribute() to 7
  assert_equals: IDL get expected (string) "7" but got (undefined) undefined
[FAIL] area.type: setAttribute() to 1.5
  assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
[FAIL] area.type: setAttribute() to "5%"
  assert_equals: IDL get expected (string) "5%" but got (undefined) undefined
[FAIL] area.type: setAttribute() to "+100"
  assert_equals: IDL get expected (string) "+100" but got (undefined) undefined
[FAIL] area.type: setAttribute() to ".5"
  assert_equals: IDL get expected (string) ".5" but got (undefined) undefined
[FAIL] area.type: setAttribute() to true
  assert_equals: IDL get expected (string) "true" but got (undefined) undefined
[FAIL] area.type: setAttribute() to false
  assert_equals: IDL get expected (string) "false" but got (undefined) undefined
[FAIL] area.type: setAttribute() to object "[object Object]"
  assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
[FAIL] area.type: setAttribute() to NaN
  assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
[FAIL] area.type: setAttribute() to Infinity
  assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
[FAIL] area.type: setAttribute() to -Infinity
  assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
[FAIL] area.type: setAttribute() to "\\0"
  assert_equals: IDL get expected (string) "\\0" but got (undefined) undefined
[FAIL] area.type: setAttribute() to null
  assert_equals: IDL get expected (string) "null" but got (undefined) undefined
[FAIL] area.type: setAttribute() to object "test-toString"
  assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
[FAIL] area.type: setAttribute() to object "test-valueOf"
  assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
[FAIL] area.type: IDL set to ""
  assert_equals: getAttribute() expected "" but got "test-valueOf"
[FAIL] area.type: IDL set to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: getAttribute() expected " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got "test-valueOf"
[FAIL] area.type: IDL set to undefined
  assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
[FAIL] area.type: IDL set to 7
  assert_equals: getAttribute() expected "7" but got "test-valueOf"
[FAIL] area.type: IDL set to 1.5
  assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
[FAIL] area.type: IDL set to "5%"
  assert_equals: getAttribute() expected "5%" but got "test-valueOf"
[FAIL] area.type: IDL set to "+100"
  assert_equals: getAttribute() expected "+100" but got "test-valueOf"
[FAIL] area.type: IDL set to ".5"
  assert_equals: getAttribute() expected ".5" but got "test-valueOf"
[FAIL] area.type: IDL set to true
  assert_equals: getAttribute() expected "true" but got "test-valueOf"
[FAIL] area.type: IDL set to false
  assert_equals: getAttribute() expected "false" but got "test-valueOf"
[FAIL] area.type: IDL set to object "[object Object]"
  assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
[FAIL] area.type: IDL set to NaN
  assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
[FAIL] area.type: IDL set to Infinity
  assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
[FAIL] area.type: IDL set to -Infinity
  assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
[FAIL] area.type: IDL set to "\\0"
  assert_equals: getAttribute() expected "\\0" but got "test-valueOf"
[FAIL] area.type: IDL set to null
  assert_equals: getAttribute() expected "null" but got "test-valueOf"
[FAIL] area.type: IDL set to object "test-toString"
  assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
[FAIL] area.type: IDL set to object "test-valueOf"
  assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
Harness: the test ran to completion.
